Add support for reading BACnet object properties.
This is temporary step in order to align naming between this facade and specification.

Attributes in spec are property identifiers.
Because wrapper defines property as a present value ther is a conflict.
There are multiple properties per object so to do not confuse past usage the 'attribute' name was picked as an temporary alternative.

Some minor adjustments in code and visitors/printers.

Signed-off-by: Łukasz Dywicki <luke@code-house.org>